+++ sdk/core/core-http/src/policies/throttlingRetryPolicy.ts
@@ -46,7 +46,10 @@ export class ThrottlingRetryPolicy extends BaseRequestPolicy {
 
   public async sendRequest(httpRequest: WebResourceLike): Promise<HttpOperationResponse> {
     return this._nextPolicy.sendRequest(httpRequest.clone()).then((response) => {
-      if (response.status !== StatusCodes.TooManyRequests) {
+      if (
+        response.status !== StatusCodes.TooManyRequests &&
+        response.status !== StatusCodes.ServiceUnavailable
+      ) {
         return response;
       } else {
         return this._handleResponse(httpRequest, response);